update: cookie_login func move to Api/cookie_login.cpp and add session_login base
