+++ html/index.html
@@ -36,6 +36,7 @@
 <h3>Bonus</h3>
 <ul>
     <li><a href="/api/cookie-login">Login (cookie)</a>
+    <li><a href="/api/session-login">Login (session)</a>
 </ul>
 
 </body>
