+++ html/login_session.html
@@ -39,9 +39,9 @@
 </head>
 
 <body>
-<h1>🍪 Login Page 🍪</h1>
+<h1>🍡 Login Page 🍡</h1>
 
-<form action="/api/cookie-login" method="post">
+<form action="/api/session-login" method="post">
     <div class="instructions">
         <p>instructions:</p>
         <ul>
