+++ srcs/HttpResponse/Api/api.cpp
@@ -12,11 +12,13 @@
 
 namespace {
 
-const char API_FORM_DATA[]      = "/api/form-data";
-const char API_RESPONSE_BODY[]  = "/api/show-response-body";
-const char API_NOW[]            = "/api/now";
-const char API_LOGIN[]          = "/api/cookie-login";
-const char API_USERPAGE[]       = "/api/cookie-userpage";
+const char API_FORM_DATA[]          = "/api/form-data";
+const char API_RESPONSE_BODY[]      = "/api/show-response-body";
+const char API_NOW[]                = "/api/now";
+const char API_COOKIE_LOGIN[]       = "/api/cookie-login";
+const char API_COOKIE_USERPAGE[]    = "/api/cookie-userpage";
+const char API_SESSION_LOGIN[]       = "/api/session-login";
+const char API_SESSION_USERPAGE[]    = "/api/session-userpage";
 
 
 std::vector<std::string> init_endpoints() {
@@ -26,8 +28,10 @@ std::vector<std::string> init_endpoints() {
     endpoints.push_back(API_FORM_DATA);
     endpoints.push_back(API_RESPONSE_BODY);
     endpoints.push_back(API_NOW);
-    endpoints.push_back(API_LOGIN);
-    endpoints.push_back(API_USERPAGE);
+    endpoints.push_back(API_COOKIE_LOGIN);
+    endpoints.push_back(API_COOKIE_USERPAGE);
+    endpoints.push_back(API_SESSION_LOGIN);
+    endpoints.push_back(API_SESSION_USERPAGE);
     return endpoints;
 }
 
@@ -57,113 +61,19 @@ StatusCode HttpResponse::response_api() {
     if (this->request_.request_target() == std::string(API_NOW)) {
         return get_now();
     }
-    if (this->request_.request_target() == std::string(API_LOGIN)) {
+    if (this->request_.request_target() == std::string(API_COOKIE_LOGIN)) {
         return get_cookie_login_page();
     }
-    if (this->request_.request_target() == std::string(API_USERPAGE)) {
+    if (this->request_.request_target() == std::string(API_COOKIE_USERPAGE)) {
         return get_cookie_user_page();
     }
+    if (this->request_.request_target() == std::string(API_SESSION_LOGIN)) {
+        return get_session_login_page();
+    }
+    if (this->request_.request_target() == std::string(API_SESSION_USERPAGE)) {
+        return get_session_user_page();
+    }
     return NotFound;
-}
-
-
-bool HttpResponse::is_logged_in_user() {
-    Result<std::map<std::string, std::string>, ProcResult> result = this->request_.get_cookie();
-    if (result.is_err()) {
-        return false;
-    }
-    std::map<std::string, std::string> cookies = result.ok_value();
-
-    std::map<std::string, std::string>::const_iterator user, email;
-    user = cookies.find("username");
-    email = cookies.find("email");
-    return user != cookies.end() && email != cookies.end();
-}
-
-
-std::string HttpResponse::get_user_name() {
-    Result<std::map<std::string, std::string>, ProcResult> result = this->request_.get_cookie();
-    if (result.is_err()) {
-        return EMPTY;
-    }
-    std::map<std::string, std::string> cookies = result.ok_value();
-
-    std::map<std::string, std::string>::const_iterator user;
-    user = cookies.find("username");
-    if (user == cookies.end()) {
-        return EMPTY;
-    }
-    return user->second;
-}
-
-
-StatusCode HttpResponse::get_cookie_user_page() {
-    if (is_logged_in_user()) {
-        const std::string head = "<!doctype html>\n"
-                                 "<html lang=\"ja\">\n"
-                                 "<head>\n"
-                                 "    <meta charset=\"UTF-8\">\n"
-                                 "    <title>Login page</title>\n"
-                                 "</head>\n"
-                                 "<body>\n"
-                                 "<h1>🍪 Login Page 🍪</h1>\n";
-
-        const std::string welcome = "<h2>Welcome, " + get_user_name() + "</h2>";
-
-        const std::string tail = "<br><br><br>\n"
-                                 "<a href=\"/\">< back to index</a>"
-                                 "</body>\n"
-                                 "</html>\n";
-
-        std::vector<unsigned char> body;
-        body.insert(body.end(), head.begin(), head.end());
-        body.insert(body.end(), welcome.begin(), welcome.end());
-        body.insert(body.end(), tail.begin(), tail.end());
-        this->body_buf_ = body;
-
-        add_content_header("html");
-        return StatusOk;
-    } else {
-        ReturnDirective redirect_to_login;
-        redirect_to_login.return_on = true;
-        redirect_to_login.code = Found;
-        redirect_to_login.text = "/login_cookie.html";
-        return get_redirect_content(redirect_to_login);
-    }
-}
-
-
-StatusCode HttpResponse::get_cookie_login_page() {
-    switch (this->request_.method()) {
-        case kGET: {
-            return get_cookie_user_page();
-        }
-        case kPOST: {
-            UrlEncodedFormData parameters = parse_urlencoded_form_data(this->body_buf_);
-
-            std::map<std::string, std::string> cookies;
-            for (UrlEncodedFormData::const_iterator itr = parameters.begin(); itr != parameters.end(); ++itr) {
-                if (itr->first == "username") {
-                    cookies["username"] = *itr->second.begin();
-                }
-                if (itr->first == "email") {
-                    cookies["email"] = *itr->second.begin();
-                }
-            }
-
-            this->cookies_ = cookies;
-            add_content_header("html");
-
-            ReturnDirective redirect_to_userpage;
-            redirect_to_userpage.return_on = true;
-            redirect_to_userpage.code = Found;
-            redirect_to_userpage.text = "/api/cookie-userpage";
-            return get_redirect_content(redirect_to_userpage);
-        }
-        default: {
-            return MethodNotAllowed;
-        }
-    }
 }
